feat(Package): better handling of internal or dependent options
Even if top level variables were set as internal, a non-internal type
was passed to the sub builds. When the top level cache was reconstructed
from the sub ones the internal type was erased. This made impossible to
hide some options to the user in ccmake and these variables were popping
in and out of ccmake because of the type change, even though the value
stayed the same